# openapi.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/v1/admin/devices/{id}/calibrate":{"put":{"tags":["admin-device-controller"],"operationId":"calibrateDevice","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CalibrateDeviceDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseDeviceDTO"}}}}}}},"/api/admin/config/timezone":{"put":{"tags":["admin-config-controller"],"operationId":"updateTimezone","requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SystemConfigDto"}}}}}}},"/api/admin/config/sample-rate":{"put":{"tags":["admin-config-controller"],"operationId":"updateSampleRate","requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SystemConfigDto"}}}}}}},"/api/admin/config/retention":{"put":{"tags":["admin-config-controller"],"operationId":"updateRetention","requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SystemConfigDto"}}}}}}},"/api/v1/auth/verification/{userVerifyToken}":{"post":{"tags":["auth-controller"],"operationId":"verifyUser","parameters":[{"name":"userVerifyToken","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/api/v1/auth/sign-up":{"post":{"tags":["auth-controller"],"operationId":"signUp","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegistrationDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/api/v1/auth/refresh-user-verification":{"post":{"tags":["auth-controller"],"operationId":"refreshUserVerification","parameters":[{"name":"username","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/api/v1/auth/refresh-token/{refreshToken}":{"post":{"tags":["auth-controller"],"operationId":"refreshToken","parameters":[{"name":"refreshToken","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseTokenResponseDTO"}}}}}}},"/api/v1/auth/logout":{"post":{"tags":["auth-controller"],"operationId":"logout","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshTokenDTO"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/v1/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseLoginResponseDTO"}}}}}}},"/api/v1/admin/users/{id}/reset-password":{"post":{"tags":["admin-user-controller"],"operationId":"resetPassword","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPasswordRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/api/v1/admin/roles":{"get":{"tags":["admin-role-controller"],"operationId":"getAllRoles","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponsePageRoleDTO"}}}}}},"post":{"tags":["admin-role-controller"],"operationId":"createRole","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRoleDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseRoleDTO"}}}}}}},"/api/v1/admin/roles/permissions":{"get":{"tags":["admin-role-controller"],"operationId":"getAllPermissions","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponsePagePermissionDTO"}}}}}},"post":{"tags":["admin-role-controller"],"operationId":"createPermission","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePermissionDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponsePermissionDTO"}}}}}}},"/api/v1/admin/devices":{"get":{"tags":["admin-device-controller"],"operationId":"getAllDevices","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponsePageDeviceDTO"}}}}}},"post":{"tags":["admin-device-controller"],"operationId":"createDevice","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateDeviceDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseDeviceDTO"}}}}}}},"/api/v1/admin/users/{id}/roles":{"patch":{"tags":["admin-user-controller"],"operationId":"updateUserRoles","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserRolesDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseUserProfileDTO"}}}}}}},"/api/v1/admin/roles/{id}/permissions":{"patch":{"tags":["admin-role-controller"],"operationId":"updateRolePermissions","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRolePermissionsDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseRoleDTO"}}}}}}},"/api/v1/users/{username}":{"get":{"tags":["user-controller"],"operationId":"getProfileByUsername","parameters":[{"name":"username","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseUserProfileDTO"}}}}}},"delete":{"tags":["user-controller"],"operationId":"deleteUser","parameters":[{"name":"username","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/admin/users":{"get":{"tags":["admin-user-controller"],"operationId":"getAllUsers","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponsePageUserProfileDTO"}}}}}}},"/api/admin/config":{"get":{"tags":["admin-config-controller"],"operationId":"getAllConfigs","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SystemConfigDto"}}}}}}}},"/api/admin/config/backup":{"get":{"tags":["admin-config-controller"],"operationId":"downloadBackup","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string","format":"binary"}}}}}}},"/api/admin/audit-log":{"get":{"tags":["audit-log-controller"],"operationId":"getAuditLogs","parameters":[{"name":"userId","in":"query","required":false,"schema":{"type":"string"}},{"name":"from","in":"query","required":false,"schema":{"type":"string","format":"date-time"}},{"name":"to","in":"query","required":false,"schema":{"type":"string","format":"date-time"}},{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageAuditLogDto"}}}}}}},"/api/v1/admin/users/{id}":{"delete":{"tags":["admin-user-controller"],"operationId":"softDeleteUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/api/v1/admin/devices/{id}":{"delete":{"tags":["admin-device-controller"],"operationId":"deleteDevice","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"CalibrateDeviceDTO":{"type":"object","properties":{"soilMoistureOffset":{"type":"number","format":"float"},"airTemperatureOffset":{"type":"number","format":"float"},"airHumidityOffset":{"type":"number","format":"float"}}},"ApiResponseDeviceDTO":{"type":"object","properties":{"status":{"type":"integer","format":"int32"},"message":{"type":"string"},"path":{"type":"string"},"timestamp":{"type":"string","format":"date-time"},"data":{"$ref":"#/components/schemas/DeviceDTO"}}},"DeviceDTO":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"userId":{"type":"string"},"username":{"type":"string"},"status":{"type":"string"},"autoWaterEnabled":{"type":"boolean"},"moistureThresholdLow":{"type":"number","format":"float"},"moistureThresholdHigh":{"type":"number","format":"float"},"soilMoistureOffset":{"type":"number","format":"float"},"airTemperatureOffset":{"type":"number","format":"float"},"airHumidityOffset":{"type":"number","format":"float"}}},"SystemConfigDto":{"type":"object","properties":{"key":{"type":"string"},"value":{"type":"string"},"updatedAt":{"type":"string","format":"date-time"}}},"ApiResponseVoid":{"type":"object","properties":{"status":{"type":"integer","format":"int32"},"message":{"type":"string"},"path":{"type":"string"},"timestamp":{"type":"string","format":"date-time"},"data":{}}},"RegistrationDTO":{"type":"object","properties":{"username":{"type":"string","minLength":1,"pattern":"^[a-z0-9]+[a-z0-9_]{3,15}$"},"password":{"type":"string","minLength":1,"pattern":"^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@$!%*?&])[A-Za-z\\d@$!%*?&]{8,20}$"},"name":{"type":"string","minLength":1},"email":{"type":"string","minLength":1,"pattern":"^[\\w-\\.]+@([\\w-]+\\.)+[\\w-]{2,4}$"}},"required":["email","name","password","username"]},"ApiResponseTokenResponseDTO":{"type":"object","properties":{"status":{"type":"integer","format":"int32"},"message":{"type":"string"},"path":{"type":"string"},"timestamp":{"type":"string","format":"date-time"},"data":{"$ref":"#/components/schemas/TokenResponseDTO"}}},"TokenResponseDTO":{"type":"object","properties":{"accessToken":{"type":"string"},"refreshToken":{"type":"string"}}},"RefreshTokenDTO":{"type":"object","properties":{"refreshToken":{"type":"string","minLength":1}},"required":["refreshToken"]},"LoginRequestDTO":{"type":"object","properties":{"username":{"type":"string","minLength":1},"password":{"type":"string","minLength":1}},"required":["password","username"]},"ApiResponseLoginResponseDTO":{"type":"object","properties":{"status":{"type":"integer","format":"int32"},"message":{"type":"string"},"path":{"type":"string"},"timestamp":{"type":"string","format":"date-time"},"data":{"$ref":"#/components/schemas/LoginResponseDTO"}}},"LoginResponseDTO":{"type":"object","properties":{"id":{"type":"string"},"username":{"type":"string"},"email":{"type":"string"},"name":{"type":"string"},"oauthId":{"type":"string"},"roles":{"type":"array","items":{"type":"string"}},"token":{"type":"string"},"refreshToken":{"type":"string"}}},"ResetPasswordRequest":{"type":"object","properties":{"newPassword":{"type":"string","minLength":1}},"required":["newPassword"]},"CreateRoleDTO":{"type":"object","properties":{"name":{"type":"string","minLength":1},"permissions":{"type":"array","items":{"type":"string"}}},"required":["name","permissions"]},"ApiResponseRoleDTO":{"type":"object","properties":{"status":{"type":"integer","format":"int32"},"message":{"type":"string"},"path":{"type":"string"},"timestamp":{"type":"string","format":"date-time"},"data":{"$ref":"#/components/schemas/RoleDTO"}}},"RoleDTO":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"permissions":{"type":"array","items":{"type":"string"}}}},"CreatePermissionDTO":{"type":"object","properties":{"name":{"type":"string","minLength":1},"description":{"type":"string"}},"required":["name"]},"ApiResponsePermissionDTO":{"type":"object","properties":{"status":{"type":"integer","format":"int32"},"message":{"type":"string"},"path":{"type":"string"},"timestamp":{"type":"string","format":"date-time"},"data":{"$ref":"#/components/schemas/PermissionDTO"}}},"PermissionDTO":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"}}},"CreateDeviceDTO":{"type":"object","properties":{"userId":{"type":"string","minLength":1},"name":{"type":"string","minLength":1}},"required":["name","userId"]},"UpdateUserRolesDTO":{"type":"object","properties":{"roles":{"type":"array","items":{"type":"string"}}},"required":["roles"]},"ApiResponseUserProfileDTO":{"type":"object","properties":{"status":{"type":"integer","format":"int32"},"message":{"type":"string"},"path":{"type":"string"},"timestamp":{"type":"string","format":"date-time"},"data":{"$ref":"#/components/schemas/UserProfileDTO"}}},"UserProfileDTO":{"type":"object","properties":{"id":{"type":"string"},"username":{"type":"string"},"email":{"type":"string"},"name":{"type":"string"},"roles":{"type":"array","items":{"type":"string"}},"createdAt":{"type":"string","format":"date-time"}}},"UpdateRolePermissionsDTO":{"type":"object","properties":{"permissions":{"type":"array","items":{"type":"string"}}},"required":["permissions"]},"Pageable":{"type":"object","properties":{"page":{"type":"integer","format":"int32","minimum":0},"size":{"type":"integer","format":"int32","minimum":1},"sort":{"type":"array","items":{"type":"string"}}}},"ApiResponsePageUserProfileDTO":{"type":"object","properties":{"status":{"type":"integer","format":"int32"},"message":{"type":"string"},"path":{"type":"string"},"timestamp":{"type":"string","format":"date-time"},"data":{"$ref":"#/components/schemas/PageUserProfileDTO"}}},"PageUserProfileDTO":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"first":{"type":"boolean"},"last":{"type":"boolean"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/UserProfileDTO"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"numberOfElements":{"type":"integer","format":"int32"},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"pageNumber":{"type":"integer","format":"int32"},"paged":{"type":"boolean"},"pageSize":{"type":"integer","format":"int32"},"unpaged":{"type":"boolean"},"offset":{"type":"integer","format":"int64"},"sort":{"$ref":"#/components/schemas/SortObject"}}},"SortObject":{"type":"object","properties":{"sorted":{"type":"boolean"},"unsorted":{"type":"boolean"},"empty":{"type":"boolean"}}},"ApiResponsePageRoleDTO":{"type":"object","properties":{"status":{"type":"integer","format":"int32"},"message":{"type":"string"},"path":{"type":"string"},"timestamp":{"type":"string","format":"date-time"},"data":{"$ref":"#/components/schemas/PageRoleDTO"}}},"PageRoleDTO":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"first":{"type":"boolean"},"last":{"type":"boolean"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/RoleDTO"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"numberOfElements":{"type":"integer","format":"int32"},"empty":{"type":"boolean"}}},"ApiResponsePagePermissionDTO":{"type":"object","properties":{"status":{"type":"integer","format":"int32"},"message":{"type":"string"},"path":{"type":"string"},"timestamp":{"type":"string","format":"date-time"},"data":{"$ref":"#/components/schemas/PagePermissionDTO"}}},"PagePermissionDTO":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"first":{"type":"boolean"},"last":{"type":"boolean"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/PermissionDTO"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"numberOfElements":{"type":"integer","format":"int32"},"empty":{"type":"boolean"}}},"ApiResponsePageDeviceDTO":{"type":"object","properties":{"status":{"type":"integer","format":"int32"},"message":{"type":"string"},"path":{"type":"string"},"timestamp":{"type":"string","format":"date-time"},"data":{"$ref":"#/components/schemas/PageDeviceDTO"}}},"PageDeviceDTO":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"first":{"type":"boolean"},"last":{"type":"boolean"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/DeviceDTO"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"numberOfElements":{"type":"integer","format":"int32"},"empty":{"type":"boolean"}}},"AuditLogDto":{"type":"object","properties":{"id":{"type":"string"},"userId":{"type":"string"},"username":{"type":"string"},"action":{"type":"string"},"targetId":{"type":"string"},"payload":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}},"PageAuditLogDto":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"first":{"type":"boolean"},"last":{"type":"boolean"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/AuditLogDto"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"numberOfElements":{"type":"integer","format":"int32"},"empty":{"type":"boolean"}}}}}}